+++ backEnd/src/main/java/com/esprit/dari/services/AdAppointmentService/interfaces/AdFavService.java
@@ -4,8 +4,8 @@ package com.esprit.dari.services.AdAppointmentService.interfaces;
 public interface AdFavService {
 
     //public void UpdateFav(AdFav estate,Long estateId);
-    public Long AddEstateFav(Long estateId,Long userId);
-    public void DeleteEstateFav(Long id);
+    public Long AddEstateFav(int estateId,Long userId);
+    public void DeleteEstateFav(long id);
    // public int countEstate();
